feat: 更新用户、论坛、订单等相关功能
- 修改UserContextUtil工具类
- 更新Order实体类
- 修改论坛评论和帖子相关Controller和Service
- 更新订单中心Controller和Service
- 删除RecommendationService推荐服务类

+++ Jsyl-springboot-background/jsyl-server/src/main/java/com/jsyl/module/forum/service/PostService.java
@@ -22,6 +22,6 @@ public interface PostService {
 
     PageResult pageQuery(PostPageQueryDTO postPageQueryDTO);
 
-    List<Post> getMyPosts(Long userId);
+    List<Post> getMyPosts(Integer userId);
 
 }
